requirements.txt
Flask
Flask-Cors
gunicorn
futures
farm-haystack
farm-haystack[ocr]